todo/api/views.py
from api.models import Todo
from .serializers import TodoSerializer
from django.views import View
from rest_framework import mixins
from rest_framework import generics


class TodoListAPIView(mixins.ListModelMixin,
                  mixins.CreateModelMixin,
                  generics.GenericAPIView):

    queryset = Todo.objects.all()
    serializer_class = TodoSerializer

    def get(self, request, *args, **kwargs):
        return self.list(request, *args, **kwargs)

    def post(self, request, *args, **kwargs):
        return self.create(request, *args, **kwargs)